move pytorch_model back to model
