+++ src/pytorch_ie/auto.py
@@ -4,10 +4,9 @@ from typing import Any, Dict, Optional
 # kept for backward compatibility
 from pie_core import AutoTaskModule
 
-from pytorch_ie.pipeline import Pipeline
-
 # kept for backward compatibility
-from pytorch_ie.pytorch_model import AutoModel
+from pytorch_ie.model import AutoModel
+from pytorch_ie.pipeline import Pipeline
 
 logger = logging.getLogger(__name__)
